+++ src/DataPersister/ProductDataPersister.php
@@ -17,13 +17,19 @@ final class ProductDataPersister implements DataPersisterInterface
     }
 
 
-    public function supports($data, ?array $context = []): bool
+    public function supports($data, array $context = []): bool
     {
+        if ($context['resource_class'] !== Product::class) {
+            return false;
+        }
         return $data instanceof Product;
     }
 
-    public function persist($data, ?array $context = [])
+    public function persist($data, array $context = [])
     {
+        if ($context['resource_class'] !== Product::class) {
+            return false;
+        }
         // Set values
         $date = new \DateTime();
         $data->setCreatedAt($date);
@@ -34,8 +40,11 @@ final class ProductDataPersister implements DataPersisterInterface
         $this->entityManager->flush();
     }
 
-    public function remove($data, ?array $context = [])
+    public function remove($data, array $context = [])
     {
+        if ($context['resource_class'] !== Product::class) {
+            return false;
+        }
         // Doctrine Remove Customer
         $this->entityManager->remove($data);
         $this->entityManager->flush();
